Change game mode and port configuration in Docker
Altered the game mode from 'survival' to 'creative' in the Docker compose YAML file. This was done to cater to user preferences for creative mode. The port mapping syntax was also updated from '- 19132:19132/udp' to '"19132:19132/udp"' to ensure compatibility with the newer version of Docker. Moreover, the volume mapping was changed to a direct path for easier data management.